+++ ISSUE_TEMPLATE.md
@@ -1,0 +1,27 @@
+<!-- Please also try to search for your issue to avoid it being closed as a duplicate. -->
+
+<!-- Change these to match your system. If you're using a version from GitHub, please mention so here along with the commit you're using. -->
+**System and QuLab version:**
+
+- macOS 10.13.3
+- Python 3.6.4
+- QuLab 0.2.1
+
+**Steps to reproduce:**
+- Code:
+```python
+
+```
+
+**Expected behavior:**
+
+
+<!-- If this is feature request, please use this section to explain the rationale behind your idea and why you believe it would be beneficial. -->
+**Actual behavior:**
+
+<!-- If you're reporting a crash, please copy the stack trace below.  If not, feel free to delete this section. -->
+
+- Crash report:
+```
+
+```
