+++ PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,6 @@
+- [ ] This change has been discussed with the author.
+- [ ] It implements / fixes issue #.
+
+---
+
+**Description:**
