+++ lab/__init__.py
@@ -5,4 +5,4 @@ from ._bootstrap import (connect_db, get_current_notebook, get_current_user,
                          save_inputCells, set_mode)
 from ._plot import make_figure_for_app, make_figures_for_App
 from ._version import __version__
-from .db._query import query
+from ._query import query
