+++ lab/_version.py
@@ -1,2 +1,2 @@
 """Define version number here and read it from setup.py automatically"""
-__version__ = "0.2.0"
+__version__ = "0.2.1"
