+++ lab/db/schema/__init__.py
@@ -1,0 +1,5 @@
+from .app import Application
+from .code_mod import CodeSnippet, Module, Notebook
+from .drv_inst import Driver, Instrument
+from .record import Record, Report, Sample, Project
+from .user import User
