+++ lab/db/update.py
@@ -1,0 +1,7 @@
+from .schema.app import saveApplication
+from .schema.code_mod import (makeUniqueCodeSnippet, makeUniqueModule,
+                              saveModule, saveModuleFile, savePackage,
+                              savePackageFile)
+from .schema.drv_inst import setInstrument, uploadDriver
+from .schema.record import newRecord
+from .schema.user import newUser
