+++ lab/db/utils.py
@@ -1,0 +1,8 @@
+import os
+
+
+def beforeSaveFile(fname):
+    '''makesure the path exists before save file'''
+    dirname = os.path.dirname(fname)
+    if not os.path.exists(dirname):
+        os.makedirs(dirname)
