+++ lab/drivers/IT6302.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
 import numpy as np
 
-from lab.device import BaseDriver
-from lab.device import QReal, QOption, QInteger, QString, QVector
+from lab.device import BaseDriver, QInteger, QOption, QReal, QString, QVector
+
 
 class Driver(BaseDriver):
-    surport_models = ['IT6302']
+    support_models = ['IT6302']
 
     quants = [
         QReal('CH1 Voltage', unit='V',
