+++ lab/drivers/Tek_AWG.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
 import numpy as np
 
-from lab.device import BaseDriver
-from lab.device import QReal, QOption, QInteger, QString, QVector
+from lab.device import BaseDriver, QInteger, QOption, QReal, QString, QVector
+
 
 class Driver(BaseDriver):
-    surport_models = ['AWG5014C', 'AWG5208']
+    support_models = ['AWG5014C', 'AWG5208']
 
     quants = [
         QReal('Sample Rate', unit='S/s',
