+++ lab/drivers/wx_AWG.py
@@ -1,10 +1,11 @@
 # -*- coding: utf-8 -*-
 import numpy as np
+
 from lab.device import BaseDriver, QInteger, QOption, QReal, QString, QVector
 
 
 class Driver(BaseDriver):
-    surport_models = ['WX1284', 'WX2184']
+    support_models = ['WX1284', 'WX2184']
 
     quants = [
         QReal('Sample Rate', unit='S/s',
